anomaly_detector/requirements.txt
quixstreams==3.23.1
python-dotenv
scikit-learn